feat(components): renaming folders
