+++ src/application.tsx
@@ -2,7 +2,7 @@ import { createBrowserRouter, RouterProvider } from 'react-router';
 import { createRoot } from 'react-dom/client';
 import { StrictMode } from 'react';
 import { SplashScreen } from './screens/splash';
-import { LoadsContextProvider } from './contexts/LoadsContext';
+import { LoadsContextProvider } from './contexts/loads-context';
 
 import './application.css';
